requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.0.0
Pillow>=9.0.0
datasets>=2.0.0
